// MS_rosterfrontend/projects/client/src/lib/view-models/client-fundinginfo.ts
export interface ClientFundingInfo {
    clientId?: number;
    ndia?: string;
    refNumber?: string;
    ageCareRecipientId?: string;
    other?: string;
    startDate?: string;
    endDate?: string;
    fundType?: number;
    fundTypeName?: string;
    amount?: string;
    claimNumber?:number;
    
}